client/pyproject.toml
[project]
name = "pkw-client"
version = "0.1.0"
description = "Client for PKWrapped."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
    "plotly==5.18.0",
    "pydantic>=2.12.5",
    "seaborn==0.13.0",
    "streamlit>=1.45.0",
]
